+++ kaithem/src/docs/changes.md
@@ -8,6 +8,7 @@ Change Log
 - :coffin: Remove kaithem.midi API
 - :coffin: Remove the image map creator util
 - :coffin: Remove kaithem.time.accuracy
+- :coffin: Remove kaithem.events.when and kaithem.events.after
 
 - :sparkles: Split off the sound stuff in a separate libary [IceMedia](https://github.com/EternityForest/icemedia) meant for easy standalone use.